+++ src/com/epochx/core/representation/SubtractFunction.java
@@ -27,8 +27,8 @@ public class SubtractFunction extends FunctionNode<Double> {
 
 	@Override
 	public Double evaluate() {
-		double c1 = getChild(0).evaluate().doubleValue();
-		double c2 = getChild(1).evaluate().doubleValue();
+		double c1 = ((Double) getChild(0).evaluate()).doubleValue();
+		double c2 = ((Double) getChild(1).evaluate()).doubleValue();
 		
 		return c1 - c2;
 	}
